Change timeout behavior so that the individual query timeout is shorter, but we requery the notary server multiple times.
This change is made to work with the newer notary server implementation
available at http://www.github.com/danwent/Perspectives-Server .  This
server performs on-demand probing by returning a 404 and expecting the
client to requery.  This approach lets individual clients make the
trade-off of how long they want to wait to see if an on-demand probe will
complete.