Build system, Unit tests - Add a perl script to validate .dtd entities when building.
Currently we check for invalid entity characters.
Removing bad characters during the build is preferable to simply detecting them during the unit tests, as some people may not run the unit tests.
Running the script as part of the build prevents us from posting a broken build, where users might have their plugin crash and not know how to fix it.
Plus you never have to remember to do it - it always runs.

Also, add a list of requirements in the README. This commit makes perl a new requirement.