Perspectives - Fix hidden bug with uninitialized variables so things actually work in SeaMonkey.
It looks like Firefox initializes the variables 'Cc' and 'Ci' by default to be references to Components.classes and Components.interfaces. But SeaMonkey does not!
In SeaMonkey those variables are undefined. Thus any SeaMonkey client running this code would hit an exception and not return any notary results, ever.
Fix that.

Also document the purpose of prompt_update while we're at it (it had an offending Cc reference).