Perspectives - Always set 'prompt_update_all_https_setting' to false
Move code that sets prompt_update_all_https_setting to a finally{} block, so we run it even if there is an error.
This way the user will see one popup max, and even if something goes wrong during the update they can continue using their browser.
If the preference wasn't set to what it "should" be, they can always change it later using the actual preferences dialog.