Unit Tests - Remove localization tests
The javascript unit tests have to be included in the extension,
so they can be installed in the browser and run.
But javascript doesn't have permission to work with the local filesystem.
These tests have to be run from a different place, as a pre-build step instead.

This update fixes the html file and javascript so unit tests can be run inside the browser.