Pers_client_policy - Fix use of quorum_size for key_weakly_seen_by_quorum()
quorum_size should actually be used in the calculation.
Now the function does what the docs say it should.
Note the new check is actually less strict -
earlier it only returned true if the key was seen by *all* notaries.
But now it is correct.

Add unit tests for this function too.