Unit tests - Fix tests for key_weakly_seen_by_quorum()
Update the tests -
the data sets used for these tests actually *are* weakly seen -
the key has been seen by at least one notary in the past X days.
These tests were incorrect before, as was the function.

The important part - that results with and without gaps are identical -
remains the same.