Perspectives - Use uri.asciiHost; Fix spinning query icon
When opening a new tab in Firefox, the uri.host is 'about:blank',
and contains some character that makes getFormattedString() unhappy.
It throws an exception.
This causes the Perspectives icon to briefly flicker to the yellow
'error' exclamation icon, which we don't want.

Simply using a string that contains a ":" doesn't trigger the exception.
Not sure what character is causing the issue, but this patch fixes it.

Also, don't start the spinning query icon unless we're actually
going to send a query (i.e. for https sites)