messages - remove 'Perspectives' header from error messages
Messages displayed as alerts are now (or should be) wrapped
in code that prepends "Perspectives: " to the start,
so it is clear where the message is coming from.
There is no longer any need to prefix alert messages.

Note that messages displayed through other means
still need to be fixed so there is a global wrap.