Perspectives - Move addToolbarButton() function to its own file
This code is based on examples from the MDN wiki
and has been released under a different license (MIT).
To separate this and denote this, and to note that
the rest of Perspectives still exists under the GPL3 license,
we separate addToolbarButton() into its own file.

Also:

- wrap both parts of the init and toolbar code in their own
'Pers_' named objects, to reduce pollution of the global namespace.

- hook up unit tests to verify these new objects

- remove debug calls