Perspectives - remove old update prompt
This prompt is from Perspectives v4.0, circe November 2011.
Five years later, it is highly likely that all clients have upgraded by now.
(Stats from AMO show only 1 user with a version < 4.0 in the past year)
If they have not, changing the setting is easy to do
and I am happy to guide anyone through it.
It's not worth keeping this old code around.